🛠️ 리팩토링: 채용공고 불필요한 컬럼 삭제

+++ src/main/java/darkoverload/itzip/feature/job/service/connect/JobInfoConnectService.java
@@ -2,17 +2,16 @@ package darkoverload.itzip.feature.job.service.connect;
 
 import darkoverload.itzip.feature.job.domain.job.JobInfo;
 import darkoverload.itzip.feature.job.domain.job.JobInfoAggregator;
-import darkoverload.itzip.feature.job.domain.job.JobInfos;
 
 import java.util.List;
 
 public interface JobInfoConnectService {
     List<JobInfo> jobInfoConnect();
 
-    Long jobInfoDelete(JobInfoAggregator jobInfoAggregator);
+    int jobInfoDelete(JobInfoAggregator jobInfoAggregator);
 
-    Long jobInfoUpdate(JobInfoAggregator jobInfoAggregator);
+    int jobInfoUpdate(JobInfoAggregator jobInfoAggregator);
 
-    Long jobInfoSave(JobInfoAggregator jobInfoAggregator);
+    int jobInfoSave(JobInfoAggregator jobInfoAggregator);
 
 }
